+++ plugins/clipboard/plugin.js
@@ -1308,7 +1308,18 @@
 
 		editor.on( 'contentDom', function() {
 			var editable = editor.editable(),
-				dropTarget = CKEDITOR.plugins.clipboard.getDropTarget( editor );
+				dropTarget = CKEDITOR.plugins.clipboard.getDropTarget( editor ),
+				top = editor.ui.space( 'top' ),
+				bottom = editor.ui.space( 'bottom' );
+
+			function preventDefaultSetDropEffectToNone( evt ) {
+				evt.data.preventDefault();
+				evt.data.$.dataTransfer.dropEffect = 'none';
+			}
+
+			// Not allowing dragging on toolbar and bottom (#12613).
+			top && top.on( 'dragover', preventDefaultSetDropEffectToNone );
+			bottom && bottom.on( 'dragover', preventDefaultSetDropEffectToNone );
 
 			// Listed on dragstart to mark internal and cross-editor drag & drop
 			// and save range and selected HTML.